optionA_etl/requirements.txt
pandas>=2.1
numpy>=1.26
requests>=2.31
python-dotenv>=1.0
pyyaml>=6.0
tqdm>=4.66
statsmodels>=0.14
pyarrow>=15.0
lxml>=5.2
